requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.4
fonttools==4.54.1
h11==0.14.0
idna==3.10
imageio==2.36.0
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
ruff==0.7.2
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
threadpoolctl==3.5.0
tifffile==2024.9.20
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
